Método Alterar Realizado

+++ Java/src/dominio/Pet.java
@@ -91,7 +91,7 @@ public class Pet {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
 }
